Minor fixes

+++ Data/Migrations/Add_reference_table.cs
@@ -17,7 +17,7 @@ namespace PI2022.Data.Migrations
                     Naziv = table.Column<string>(type: "nvarchar(max)", nullable: false),
                     Kategorija = table.Column<string>(type: "nvarchar(max)", nullable: false),
                     Opis = table.Column<string>(type: "nvarchar(max)", nullable: false),
-                    CijenaReferentneOpreme = table.Column<int>(type: "int", nullable: false)
+                    CijenaReferentneOpreme = table.Column<Double>(type: "float", nullable: false)
                 },
                 constraints: table =>
                 {
